web/requirements.txt
acme==4.0.0
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
bcrypt==4.3.0
blinker==1.9.0
boto3==1.38.32
botocore==1.38.32
cachetools==5.5.2
certbot==4.0.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
ConfigArgParse==1.7.1
configobj==5.0.9
cryptography==45.0.3
distro==1.9.0
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-RESTful==0.3.10
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.1
grpcio-status==1.71.0
gspread==6.2.1
gspread-dataframe==4.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
josepy==2.0.0
lxml==5.4.0
MarkupSafe==3.0.2
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.84.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
parsedatetime==2.6
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
pymongo==4.13.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
redis==6.2.0
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
s3transfer==0.13.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.22.0
zipstream-new==1.1.8